add function typedefs for AEAD encrypt/decrypt
